Add support for the paste and match style command
