+++ src/libs/Clipboard/index.native.js
@@ -1,3 +1,15 @@
 import Clipboard from '@react-native-community/clipboard';
 
-export default Clipboard;
+/**
+ * Writes the content as text on native.
+ * @param {Object} content Types to write.
+ * @param {String} content.markdown MD representation of the content.
+ */
+const writeTypes = (content) => {
+    Clipboard.setString(content.markdown);
+};
+
+export default {
+    ...Clipboard,
+    writeTypes,
+};
