+++ src/pages/home/report/ContextMenu/ContextMenuUtils/index.js
@@ -4,12 +4,12 @@
  * The popover description will be an empty string on anything but mobile web
  * because we show link in tooltip instead of popover on web
  *
- * @param {String} selection
+ * @param {Object} selection
  * @param {Boolean} isSmallScreenWidth
  * @returns {String}
  */
 function getPopoverDescription(selection, isSmallScreenWidth) {
-    return isSmallScreenWidth ? selection : '';
+    return isSmallScreenWidth && selection ? selection.text : '';
 }
 
 export {
