+++ src/pages/home/report/ContextMenu/ContextMenuUtils/index.native.js
@@ -1,13 +1,15 @@
 /* eslint-disable import/prefer-default-export */
 
+import _ from 'lodash';
+
 /**
  * Always show popover description on native platforms
  *
- * @param {String} selection
+ * @param {Object} selection
  * @returns {String}
  */
 function getPopoverDescription(selection) {
-    return selection;
+    return _.get(selection, 'text', '');
 }
 
 export {
