+++ src/pages/home/report/ContextMenu/ReportActionContextMenu.js
@@ -7,7 +7,7 @@ const contextMenuRef = React.createRef();
  *
  * @param {string} type - the context menu type to display [EMAIL, LINK, REPORT_ACTION]
  * @param {Object} [event] - A press event.
- * @param {string} [selection] - A copy text.
+ * @param {Object} [selection] - Copied content.
  * @param {Element} contextMenuAnchor - popoverAnchor
  * @param {Number} reportID - Active Report Id
  * @param {Object} reportAction - ReportAction for ContextMenu
